+++ test/ModuleInterface/ModuleCache/Inputs/sdk-build-ver.1.plist
@@ -1,0 +1,6 @@
+<plist version="1.0">
+<dict>
+	<key>ProductBuildVersion</key>
+	<string>11111</string>
+</dict>
+</plist>
